chore: rename internal
