+++ packages/react-server/src/runtime-server.ts
@@ -1,2 +1,2 @@
-export { registerClientReference } from "./features/use-client/react-server";
-export { registerServerReference } from "./features/server-action/react-server";
+export { registerClientReference } from "./features/use-client/server";
+export { registerServerReference } from "./features/server-action/server";
